fix(api): T3 — combinar integration tests cuentas_demo en un solo file
CI run de PR #337 falló porque los 2 integration test files separados
(seed-demo-second-cold-start + seed-demo-third-cold-start) observaron
cross-file state pollution: cada file recrea `handle.pool` en su propio
beforeAll, pero la BD subyacente es la misma, y singleFork de vitest no
garantiza file-level isolation completo (la primera file's INSERTs
sobreviven a la segunda's TRUNCATE bajo MVCC + connection state diferente).

Solución: combinar los 2 escenarios en un único file
seed-demo-cuentas-idempotency.integration.test.ts con beforeEach común
(DELETE + assertion count=0). Misma describe block → vitest garantiza
serial execution + state cleanup verificable per test.

Cambios:
- New: apps/api/test/integration/seed-demo-cuentas-idempotency.integration.test.ts
  (4 tests: second-cold-start hit; persona solo disabled; N=3 cold-starts;
  N=5 invocaciones consecutivas).
- Removed: seed-demo-second-cold-start.integration.test.ts.
- Removed: seed-demo-third-cold-start.integration.test.ts.
- DELETE en lugar de TRUNCATE: MVCC-friendly, evita DDL lock contention.

Closes regression CI run 26406979453 (T3 PR #337).

Tests: typecheck green; biome lint clean; integration tests gating en
T0 job CI verificará idempotency contra postgres real serializado.
